feat(csi): restructure the CSI as different mr executable (#601)
* Restructure the CSI as different executable


* Move csi under /cmd


---------
