+++ .gitignore
@@ -43,3 +43,9 @@ readme.txt
 
 #OSX files
 **/.DS_Store
+
+# CSI executable
+mr-storage-initializer
+
+# KServe manifests
+istio-*
