+++ test/scripts/install_kustomize.sh
